revert: loong64: 降级至1.82版本

+++ src/doc/book/listings/ch11-writing-automated-tests/listing-11-07/src/lib.rs
@@ -1,4 +1,4 @@
-pub fn add_two(a: u64) -> u64 {
+pub fn add_two(a: usize) -> usize {
     a + 2
 }
