+++ src/doc/book/listings/ch11-writing-automated-tests/listing-11-12/src/lib.rs
@@ -1,8 +1,8 @@
-pub fn add_two(a: u64) -> u64 {
+pub fn add_two(a: usize) -> usize {
     internal_adder(a, 2)
 }
 
-fn internal_adder(left: u64, right: u64) -> u64 {
+fn internal_adder(left: usize, right: usize) -> usize {
     left + right
 }
