+++ src/doc/book/listings/ch11-writing-automated-tests/no-listing-10-result-in-tests/src/lib.rs
@@ -1,12 +1,12 @@
-pub fn add(left: u64, right: u64) -> u64 {
+pub fn add(left: usize, right: usize) -> usize {
     left + right
 }
 
-// ANCHOR: here
 #[cfg(test)]
 mod tests {
     use super::*;
 
+    // ANCHOR: here
     #[test]
     fn it_works() -> Result<(), String> {
         let result = add(2, 2);
@@ -17,5 +17,5 @@ mod tests {
             Err(String::from("two plus two does not equal four"))
         }
     }
+    // ANCHOR_END: here
 }
-// ANCHOR_END: here
